pushing some fixes from may 27th

+++ c4data/lisaData/LisaAnaData.h
@@ -49,8 +49,8 @@ class LisaAnaItem : public TObject
         LisaAnaItem();
 
         void SetAll(uint64_t br,
-                    uint16_t id,
-                    uint64_t rn,
+                    uint32_t id,
+                    uint32_t rn,
                     int b_id,
                     uint64_t ev_time,
                     int ch_id,
@@ -69,8 +69,8 @@ class LisaAnaItem : public TObject
         void Reset();
 
         uint64_t Get_br_t() const;
-        uint16_t Get_br_id() const;
-        uint64_t Get_run_number() const;
+        uint32_t Get_br_id() const;
+        uint32_t Get_run_number() const;
         int Get_board_id() const;
         uint64_t Get_board_event_time() const;
         int Get_channel_id() const;
@@ -92,8 +92,8 @@ class LisaAnaItem : public TObject
     private:
 
         uint64_t br_t;
-        uint16_t br_id;
-        uint64_t run_number;
+        uint32_t br_id;
+        uint32_t run_number;
         int board_id;
         uint64_t board_event_time;
         int channel_id;
@@ -117,12 +117,12 @@ inline uint64_t LisaAnaItem::Get_br_t() const
     return br_t;
 }
 
-inline uint16_t LisaAnaItem::Get_br_id() const
+inline uint32_t LisaAnaItem::Get_br_id() const
 {
     return br_id;
 }
 
-inline uint64_t LisaAnaItem::Get_run_number() const
+inline uint32_t LisaAnaItem::Get_run_number() const
 {
     return run_number;
 }
